refactor(emit): split recovered statement helpers
