+++ crates/tsz-emitter/src/emitter/statements/mod.rs
@@ -4,6 +4,7 @@ mod core;
 mod expression_statement_helpers;
 mod recovered_expression_statement_helpers;
 mod recovered_generated_member_helpers;
+mod recovered_variable_statement;
 mod static_block_await_recovery;
 mod try_statement;
 mod variable_statement_helpers;
